Remove distro-specific stuff and various hacks
Debian certainly doesn't use the debian.in, I'm guessing the same goes for
the rest. Will consider adding debian/ and redhat/ directories if they look
somewhat more complete and sensible.

For the subst and update-version stuff... It's either redundant, hackish or
plain Wrong. Will find a better solution if it's even needed.